@@ -2,9 +2,12 @@
   "name": "@kontent-ai/sample-app-vue",
   "version": "1.0.1",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "run-p type-check build-only",
+    "preview": "vite preview --port 4173",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore"
   },
   "lint-staged": {
     "*.{js,vue}": [
@@ -16,28 +19,39 @@
     "pre-commit": "lint-staged"
   },
   "dependencies": {
-    "@intlify/vue-i18n-loader": "^1.0.0",
-    "@kontent-ai/delivery-sdk": "^12.0.2",
-    "dateformat": "^3.0.3",
-    "lodash": "^4.17.19",
-    "qs": "^6.9.4",
-    "rxjs": "^6.5.5",
+    "@intlify/vue-i18n-loader": "^4.2.0",
+    "@kontent-ai/delivery-sdk": "^12.2.0",
+    "dateformat": "^5.0.3",
+    "lodash": "^4.17.21",
+    "qs": "^6.11.0",
+    "rxjs": "^7.5.6",
     "twitter-widgets": "^2.0.0",
-    "universal-cookie": "^4.0.3",
-    "validator": "^13.0.0",
-    "vue": "^2.6.11",
-    "vue-i18n": "^8.18.1",
-    "vue-router": "^3.3.2",
-    "vue-scrollto": "^2.18.1",
+    "universal-cookie": "^4.0.4",
+    "validator": "^13.7.0",
+    "vue": "^3.2.39",
+    "vue-i18n": "^9.2.2",
+    "vue-router": "^4.1.5",
+    "vue-scrollto": "^2.20.0",
     "vue2-google-maps": "^0.10.7"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.8.0",
-    "@vue/cli-plugin-eslint": "^3.12.1",
-    "@vue/cli-service": "^4.4.1",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "@vue/eslint-config-typescript": "^11.0.0",
+    "@vue/tsconfig": "^0.1.3",
+    "@vitejs/plugin-vue": "^3.0.3",
+    "npm-run-all": "^4.1.5",
     "deepmerge": "^4.2.2",
-    "lint-staged": "^10.2.7",
-    "vue-template-compiler": "^2.6.11"
+    "lint-staged": "^13.0.3",
+    "vue-template-compiler": "^2.7.10",
+    "eslint": "^8.22.0",
+    "eslint-plugin-vue": "^9.3.0",
+    "prettier": "^2.7.1",
+    "typescript": "~4.8.3",
+    "vite": "^3.0.9",
+    "vue-tsc": "^0.40.7"
   },
   "babel": {
     "presets": [
